copyonwrite: fix crash in test3()
test() performs the test writing to the source memory which is the
destination memory of the other source memory.

This is the fix of the regression of commit
55bb5e685ab7453de5b289e2b8b7582db03bc101. That commit hid this failure.
Because a stack is always writable without modifying its access flag on
purpose.

However, WRITE permission is removed from all registered destination
memory. When this memory becomes the source memory of the other
destination memory, its permission flags would not have WRITE flag.

In exception handler, it determines this memory did not have WRITE
permission orignally. Write acess exception is not handled properly.

After all, crash occurs.

    modified:   memory/copyonwrite.c